backend/requirements.txt
# testes 1 e 2
requests
beautifulsoup4
pandas
PyPDF2
tabula-py
zipfile36
JPype1 # Para rodar o tabula-py

# teste 3
psycopg2-binary
sqlalchemy 

# teste 4
fastapi
uvicorn
pydantic